Make eigenstep.sage a function
Rather than having a sage script with a hardcoded eigensteps matrix use
a function ConstructFrameFromEigensteps(A), where A is a matrix in the
desired format.

More work has to be done to make the function accept standard eigenstep
matrices (i.e without the all-zero-line) and some error handling (i.e.
if the matrix doesn't represent a valid sequence of eigensteps).